requirements.txt
streamlit==1.32.0
chromadb==0.4.24
google-generativeai==0.4.0
python-dotenv==1.0.1